Slim down billing event payloads to only id, municipalityId and eventType — billing fetches contract details itself via GET
