+++ src/integration-test/resources/OutboxIT/__files/test06_dispatchSendsDeletedEventToBilling/mappings/billing-contract-deleted.json
@@ -1,7 +1,7 @@
 {
 	"request": {
 		"method": "POST",
-		"url": "/api-billing-data-collector/1984/contracts/deleted"
+		"url": "/api-billing-data-collector/1984/CONTRACTS/events"
 	},
 	"response": {
 		"status": 200
