+++ src/main/java/se/sundsvall/contract/integration/billingdatacollector/BillingDataCollectorClient.java
@@ -6,10 +6,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import se.sundsvall.contract.integration.billingdatacollector.configuration.BillingDataCollectorConfiguration;
-import se.sundsvall.contract.integration.billingdatacollector.event.ContractCreatedEvent;
-import se.sundsvall.contract.integration.billingdatacollector.event.ContractDeletedEvent;
-import se.sundsvall.contract.integration.billingdatacollector.event.ContractTerminatedEvent;
-import se.sundsvall.contract.integration.billingdatacollector.event.ContractUpdatedEvent;
+import se.sundsvall.contract.integration.billingdatacollector.event.BillingEvent;
 
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static se.sundsvall.contract.integration.billingdatacollector.configuration.BillingDataCollectorConfiguration.CLIENT_ID;
@@ -26,46 +23,15 @@ import static se.sundsvall.contract.integration.billingdatacollector.configurati
 public interface BillingDataCollectorClient {
 
 	/**
-	 * Notify billing that a contract has been created
+	 * Notify billing about a contract event. Billing fetches the contract data itself
+	 * via GET /{municipalityId}/contracts/{id} using the provided id and municipalityId.
 	 *
-	 * @param municipalityId the municipality id of the municipality that owns the contract
-	 * @param event          the contract created event payload
+	 * @param municipalityId the municipality id
+	 * @param event          the billing event payload
 	 */
-	@PostMapping(path = "/{municipalityId}/contracts/created", consumes = APPLICATION_JSON_VALUE)
-	void contractCreated(
+	@PostMapping(path = "/{municipalityId}/{source}/events", consumes = APPLICATION_JSON_VALUE)
+	void sendEvent(
 		@PathVariable final String municipalityId,
-		@RequestBody final ContractCreatedEvent event);
-
-	/**
-	 * Notify billing that a contract has been updated
-	 *
-	 * @param municipalityId the municipality id of the municipality that owns the contract
-	 * @param event          the contract updated event payload
-	 */
-	@PostMapping(path = "/{municipalityId}/contracts/updated", consumes = APPLICATION_JSON_VALUE)
-	void contractUpdated(
-		@PathVariable final String municipalityId,
-		@RequestBody final ContractUpdatedEvent event);
-
-	/**
-	 * Notify billing that a contract has been deleted
-	 *
-	 * @param municipalityId the municipality id of the municipality that owns the contract
-	 * @param event          the contract deleted event payload
-	 */
-	@PostMapping(path = "/{municipalityId}/contracts/deleted", consumes = APPLICATION_JSON_VALUE)
-	void contractDeleted(
-		@PathVariable final String municipalityId,
-		@RequestBody final ContractDeletedEvent event);
-
-	/**
-	 * Notify billing that a contract has been terminated
-	 *
-	 * @param municipalityId the municipality id of the municipality that owns the contract
-	 * @param event          the termination event payload
-	 */
-	@PostMapping(path = "/{municipalityId}/contracts/terminated", consumes = APPLICATION_JSON_VALUE)
-	void contractTerminated(
-		@PathVariable final String municipalityId,
-		@RequestBody final ContractTerminatedEvent event);
+		@PathVariable final BillingSource source,
+		@RequestBody final BillingEvent event);
 }
