+++ src/main/java/se/sundsvall/contract/integration/billingdatacollector/BillingSource.java
@@ -1,0 +1,5 @@
+package se.sundsvall.contract.integration.billingdatacollector;
+
+public enum BillingSource {
+	CONTRACTS
+}
