+++ src/main/java/se/sundsvall/contract/integration/billingdatacollector/HttpBillingEventPublisher.java
@@ -2,10 +2,6 @@ package se.sundsvall.contract.integration.billingdatacollector;
 
 import org.springframework.stereotype.Component;
 import se.sundsvall.contract.integration.billingdatacollector.event.BillingEvent;
-import se.sundsvall.contract.integration.billingdatacollector.event.ContractCreatedEvent;
-import se.sundsvall.contract.integration.billingdatacollector.event.ContractDeletedEvent;
-import se.sundsvall.contract.integration.billingdatacollector.event.ContractTerminatedEvent;
-import se.sundsvall.contract.integration.billingdatacollector.event.ContractUpdatedEvent;
 
 @Component
 public class HttpBillingEventPublisher implements BillingEventPublisher {
@@ -18,11 +14,6 @@ public class HttpBillingEventPublisher implements BillingEventPublisher {
 
 	@Override
 	public void publish(final BillingEvent event) {
-		switch (event) {
-			case ContractCreatedEvent e -> client.contractCreated(e.municipalityId(), e);
-			case ContractUpdatedEvent e -> client.contractUpdated(e.municipalityId(), e);
-			case ContractDeletedEvent e -> client.contractDeleted(e.municipalityId(), e);
-			case ContractTerminatedEvent e -> client.contractTerminated(e.municipalityId(), e);
-		}
+		client.sendEvent(event.municipalityId(), BillingSource.CONTRACTS, event);
 	}
 }
