+++ src/main/java/se/sundsvall/contract/integration/billingdatacollector/event/BillingEvent.java
@@ -2,7 +2,7 @@ package se.sundsvall.contract.integration.billingdatacollector.event;
 
 public sealed interface BillingEvent permits ContractCreatedEvent, ContractUpdatedEvent, ContractDeletedEvent, ContractTerminatedEvent {
 
-	String contractId();
+	String id();
 
 	String municipalityId();
