+++ src/test/java/se/sundsvall/contract/integration/billingdatacollector/HttpBillingEventPublisherTest.java
@@ -1,6 +1,5 @@
 package se.sundsvall.contract.integration.billingdatacollector;
 
-import java.time.LocalDate;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -11,11 +10,6 @@ import se.sundsvall.contract.integration.billingdatacollector.event.ContractCrea
 import se.sundsvall.contract.integration.billingdatacollector.event.ContractDeletedEvent;
 import se.sundsvall.contract.integration.billingdatacollector.event.ContractTerminatedEvent;
 import se.sundsvall.contract.integration.billingdatacollector.event.ContractUpdatedEvent;
-import se.sundsvall.contract.model.enums.ContractType;
-import se.sundsvall.contract.model.enums.IntervalType;
-import se.sundsvall.contract.model.enums.InvoicedIn;
-import se.sundsvall.contract.model.enums.LeaseType;
-import se.sundsvall.contract.model.enums.Status;
 
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
@@ -25,8 +19,6 @@ class HttpBillingEventPublisherTest {
 
 	private static final String MUNICIPALITY_ID = "2281";
 	private static final String CONTRACT_ID = "CONTRACT-1";
-	private static final LocalDate START_DATE = LocalDate.of(2020, 1, 1);
-	private static final LocalDate END_DATE = LocalDate.of(2026, 6, 30);
 
 	@Mock
 	private BillingDataCollectorClient clientMock;
@@ -41,51 +33,29 @@ class HttpBillingEventPublisherTest {
 
 	@Test
 	void publishContractCreatedEvent() {
-		// Arrange
-		final var event = new ContractCreatedEvent(CONTRACT_ID, MUNICIPALITY_ID, ContractType.LEASE_AGREEMENT, Status.ACTIVE,
-			START_DATE, END_DATE, null, null, InvoicedIn.ADVANCE, IntervalType.QUARTERLY, LeaseType.LAND_LEASE_RESIDENTIAL);
-
-		// Act
+		final var event = new ContractCreatedEvent(CONTRACT_ID, MUNICIPALITY_ID);
 		publisher.publish(event);
-
-		// Assert
-		verify(clientMock).contractCreated(MUNICIPALITY_ID, event);
+		verify(clientMock).sendEvent(MUNICIPALITY_ID, BillingSource.CONTRACTS, event);
 	}
 
 	@Test
 	void publishContractUpdatedEvent() {
-		// Arrange
-		final var event = new ContractUpdatedEvent(CONTRACT_ID, MUNICIPALITY_ID, ContractType.LEASE_AGREEMENT, Status.ACTIVE,
-			START_DATE, END_DATE, null, null, InvoicedIn.ARREARS, IntervalType.YEARLY, LeaseType.LAND_LEASE_RESIDENTIAL);
-
-		// Act
+		final var event = new ContractUpdatedEvent(CONTRACT_ID, MUNICIPALITY_ID);
 		publisher.publish(event);
-
-		// Assert
-		verify(clientMock).contractUpdated(MUNICIPALITY_ID, event);
+		verify(clientMock).sendEvent(MUNICIPALITY_ID, BillingSource.CONTRACTS, event);
 	}
 
 	@Test
 	void publishContractDeletedEvent() {
-		// Arrange
 		final var event = new ContractDeletedEvent(CONTRACT_ID, MUNICIPALITY_ID);
-
-		// Act
 		publisher.publish(event);
-
-		// Assert
-		verify(clientMock).contractDeleted(MUNICIPALITY_ID, event);
+		verify(clientMock).sendEvent(MUNICIPALITY_ID, BillingSource.CONTRACTS, event);
 	}
 
 	@Test
 	void publishContractTerminatedEvent() {
-		// Arrange
-		final var event = new ContractTerminatedEvent(CONTRACT_ID, MUNICIPALITY_ID, END_DATE, InvoicedIn.ADVANCE, IntervalType.QUARTERLY);
-
-		// Act
+		final var event = new ContractTerminatedEvent(CONTRACT_ID, MUNICIPALITY_ID);
 		publisher.publish(event);
-
-		// Assert
-		verify(clientMock).contractTerminated(MUNICIPALITY_ID, event);
+		verify(clientMock).sendEvent(MUNICIPALITY_ID, BillingSource.CONTRACTS, event);
 	}
 }
