+++ src/test/resources/api/openapi.yaml
@@ -5,7 +5,7 @@ info:
   license:
     name: MIT License
     url: https://opensource.org/licenses/MIT
-  version: "7.1"
+  version: "7.0"
 servers:
   - url: http://localhost:52742
     description: Generated server url
